ajustes

+++ zipFiles.py
@@ -1,4 +1,4 @@
-import os, shutil, time
+import os, shutil, time, sys
 from datetime import datetime
 
 pathRoot = os.getcwd()
@@ -14,16 +14,29 @@ def generateFolderDefault():
     return os.path.normpath( os.path.join( pathRoot, folderName ) )
 
 def checkFolderNone( dirpath ):
+    if not os.path.exists( dirpath ):
+        print('Não existe a pasta files')
+        sys.exit()
+
     dirFiles = os.listdir( dirpath )
     for dirFile in dirFiles:
         dirNext = not os.listdir( os.path.join( dirpath, dirFile ) )
         if dirNext == True:
             os.system( 'rmdir "%s"' % os.path.join( dirpath, dirFile ) )
+        else:
+            for fileName in os.listdir( os.path.join( dirpath, dirFile ) ):
+                if os.path.isfile( os.path.join( os.path.join( dirpath, dirFile ), 'file.json' ) ) and fileName.endswith(".png"):
+                    #print(dirFile)
+                    break
+                else:
+                    #remove as pastas sem json ou sem imagem
+                    removeFolderFile( os.path.join( dirpath, dirFile ) )
+                    break
+
 
 
 def removeFolderFile( dirpath ):
    shutil.rmtree( dirpath )
-
 
 def zipAll( source, destination ):
     generateFolderDefault()
